meta-openembedded: nginx: add stable version 1.8.1 as alternative to the mainline version 1.9.x
The 1.8 branch is the current stable branch of nginx.
This means the branch doesn't get new features, but is still supported with bugfixes.

Depending on the use case it is more suitable to use on an embedded device
than the 1.9 branch which adds new features with every release.

(From meta-openembedded rev: 6605e7f52f5880c0dd4eb87e514d42827ee58050)

Signed-off-by: Pascal Bach <pascal.bach@siemens.com>