u-boot.inc: Add sub-dir support for SPL_BINARY
Add support for the SPL_BINARY variable to handle sub directories. In
some cases the SPL binary that needs to be deployed is only built to the
spl/ directory in U-Boot. So that a sub directory can be specified in
the SPL_BINARY variable, handle the case so that the deploy code uses
the basename of the path specified in SPL_BINARY.

(From OE-core rev: eb90d1c8fc7b82ca2593185930b3bf175f40ae13)

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>